Where the magic happens
Removed the english based checking system that was in place to find the gems to be used.  Not everyone in the community runs the client with the english language.  Quite the oversight on my part, but its fixed now and uses the base identifier keys to check for the items, which is a check independent of language.